KBucket test moved
